requirements.txt
aiohttp
beautifulsoup4
PyYAML
python-dotenv
requests
streamlit
pandas
plotly